+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
   - Fix SSL generation script and strengthen from 2048 bit to 4096 bit RSA ([#234](https://github.com/kizniche/mycodo/issues/234))
 
+### Miscellaneous
+
+  - New cleaner Timer page style
+
 ## 5.0.21 (2017-04-02)
 
 ### Bugfixes